requirements.txt
streamlit
tensorflow
pillow
numpy
huggingface-hub

